requirements.txt
coverage
fastapi
httpx
passlib[bcrypt]
pylint
pytest
python-jose[cryptography]
python-multipart
uvicorn[standard]
